Shorten agent code
Move code out of agent that is general to dedicated files, no logic
changes just moving.
